Modularize chat interface
