+++ clients/metorial-dashboard/src/gen/src/mt_2025_01_01_dashboard/resources/dashboard/instance/skills/list.ts
@@ -8,6 +8,11 @@ export type DashboardInstanceSkillsListOutput = {
     slug: string;
     name: string;
     description: string | null;
+    clientName: string;
+    clientDescription: string | null;
+    clientMetadata: Record<string, any> | null;
+    license: string | null;
+    compatibility: string | null;
     metadata: Record<string, any>;
     storeId: string;
     hierarchy: {
@@ -69,6 +74,20 @@ export let mapDashboardInstanceSkillsListOutput =
           slug: mtMap.objectField('slug', mtMap.passthrough()),
           name: mtMap.objectField('name', mtMap.passthrough()),
           description: mtMap.objectField('description', mtMap.passthrough()),
+          clientName: mtMap.objectField('client_name', mtMap.passthrough()),
+          clientDescription: mtMap.objectField(
+            'client_description',
+            mtMap.passthrough()
+          ),
+          clientMetadata: mtMap.objectField(
+            'client_metadata',
+            mtMap.passthrough()
+          ),
+          license: mtMap.objectField('license', mtMap.passthrough()),
+          compatibility: mtMap.objectField(
+            'compatibility',
+            mtMap.passthrough()
+          ),
           metadata: mtMap.objectField('metadata', mtMap.passthrough()),
           storeId: mtMap.objectField('store_id', mtMap.passthrough()),
           hierarchy: mtMap.objectField(
