+++ clients/metorial-dashboard/src/gen/src/mt_2025_01_01_dashboard/resources/management/instance/skill-templates/create.ts
@@ -126,12 +126,14 @@ export type ManagementInstanceSkillTemplatesCreateBody = {
   name: string;
   description?: string | undefined;
   metadata?: Record<string, any> | undefined;
+  fromSkillId?: string | undefined;
 };
 
 export let mapManagementInstanceSkillTemplatesCreateBody =
   mtMap.object<ManagementInstanceSkillTemplatesCreateBody>({
     name: mtMap.objectField('name', mtMap.passthrough()),
     description: mtMap.objectField('description', mtMap.passthrough()),
-    metadata: mtMap.objectField('metadata', mtMap.passthrough())
+    metadata: mtMap.objectField('metadata', mtMap.passthrough()),
+    fromSkillId: mtMap.objectField('from_skill_Id', mtMap.passthrough())
   });
