+++ clients/metorial-dashboard/src/gen/src/mt_2025_01_01_dashboard/resources/management/instance/skills/duplicate.ts
@@ -7,6 +7,11 @@ export type ManagementInstanceSkillsDuplicateOutput = {
   slug: string;
   name: string;
   description: string | null;
+  clientName: string;
+  clientDescription: string | null;
+  clientMetadata: Record<string, any> | null;
+  license: string | null;
+  compatibility: string | null;
   metadata: Record<string, any>;
   storeId: string;
   hierarchy: {
@@ -62,6 +67,14 @@ export let mapManagementInstanceSkillsDuplicateOutput =
     slug: mtMap.objectField('slug', mtMap.passthrough()),
     name: mtMap.objectField('name', mtMap.passthrough()),
     description: mtMap.objectField('description', mtMap.passthrough()),
+    clientName: mtMap.objectField('client_name', mtMap.passthrough()),
+    clientDescription: mtMap.objectField(
+      'client_description',
+      mtMap.passthrough()
+    ),
+    clientMetadata: mtMap.objectField('client_metadata', mtMap.passthrough()),
+    license: mtMap.objectField('license', mtMap.passthrough()),
+    compatibility: mtMap.objectField('compatibility', mtMap.passthrough()),
     metadata: mtMap.objectField('metadata', mtMap.passthrough()),
     storeId: mtMap.objectField('store_id', mtMap.passthrough()),
     hierarchy: mtMap.objectField(
@@ -146,6 +159,11 @@ export let mapManagementInstanceSkillsDuplicateOutput =
 export type ManagementInstanceSkillsDuplicateBody = {
   name: string;
   description?: string | undefined;
+  clientName?: string | undefined;
+  clientDescription?: string | undefined;
+  license?: string | undefined;
+  compatibility?: string | undefined;
+  clientMetadata?: Record<string, any> | undefined;
   metadata?: Record<string, any> | undefined;
 };
 
@@ -153,6 +171,14 @@ export let mapManagementInstanceSkillsDuplicateBody =
   mtMap.object<ManagementInstanceSkillsDuplicateBody>({
     name: mtMap.objectField('name', mtMap.passthrough()),
     description: mtMap.objectField('description', mtMap.passthrough()),
+    clientName: mtMap.objectField('client_name', mtMap.passthrough()),
+    clientDescription: mtMap.objectField(
+      'client_description',
+      mtMap.passthrough()
+    ),
+    license: mtMap.objectField('license', mtMap.passthrough()),
+    compatibility: mtMap.objectField('compatibility', mtMap.passthrough()),
+    clientMetadata: mtMap.objectField('client_metadata', mtMap.passthrough()),
     metadata: mtMap.objectField('metadata', mtMap.passthrough())
   });
