+++ clients/metorial-dashboard/src/gen/src/mt_2025_01_01_dashboard/resources/skills/create.ts
@@ -7,6 +7,11 @@ export type SkillsCreateOutput = {
   slug: string;
   name: string;
   description: string | null;
+  clientName: string;
+  clientDescription: string | null;
+  clientMetadata: Record<string, any> | null;
+  license: string | null;
+  compatibility: string | null;
   metadata: Record<string, any>;
   storeId: string;
   hierarchy: {
@@ -61,6 +66,14 @@ export let mapSkillsCreateOutput = mtMap.object<SkillsCreateOutput>({
   slug: mtMap.objectField('slug', mtMap.passthrough()),
   name: mtMap.objectField('name', mtMap.passthrough()),
   description: mtMap.objectField('description', mtMap.passthrough()),
+  clientName: mtMap.objectField('client_name', mtMap.passthrough()),
+  clientDescription: mtMap.objectField(
+    'client_description',
+    mtMap.passthrough()
+  ),
+  clientMetadata: mtMap.objectField('client_metadata', mtMap.passthrough()),
+  license: mtMap.objectField('license', mtMap.passthrough()),
+  compatibility: mtMap.objectField('compatibility', mtMap.passthrough()),
   metadata: mtMap.objectField('metadata', mtMap.passthrough()),
   storeId: mtMap.objectField('store_id', mtMap.passthrough()),
   hierarchy: mtMap.objectField(
@@ -143,11 +156,24 @@ export type SkillsCreateBody = {
   name: string;
   description?: string | undefined;
   metadata?: Record<string, any> | undefined;
+  clientName?: string | undefined;
+  clientDescription?: string | undefined;
+  license?: string | undefined;
+  compatibility?: string | undefined;
+  clientMetadata?: Record<string, any> | undefined;
 };
 
 export let mapSkillsCreateBody = mtMap.object<SkillsCreateBody>({
   name: mtMap.objectField('name', mtMap.passthrough()),
   description: mtMap.objectField('description', mtMap.passthrough()),
-  metadata: mtMap.objectField('metadata', mtMap.passthrough())
+  metadata: mtMap.objectField('metadata', mtMap.passthrough()),
+  clientName: mtMap.objectField('client_name', mtMap.passthrough()),
+  clientDescription: mtMap.objectField(
+    'client_description',
+    mtMap.passthrough()
+  ),
+  license: mtMap.objectField('license', mtMap.passthrough()),
+  compatibility: mtMap.objectField('compatibility', mtMap.passthrough()),
+  clientMetadata: mtMap.objectField('client_metadata', mtMap.passthrough())
 });
