+++ src/frontend/scenes/assistant/src/components/textShimmer.tsx
