+++ src/frontend/scenes/assistant/src/components/tools/searchTool.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import type { AssistantLiveStateItem } from '@metorial/state';
 import { Error } from '@metorial/ui';
 import styled from 'styled-components';
