+++ src/frontend/scenes/assistant/src/scenes/index.ts
@@ -1,0 +1,2 @@
+export * from './conversation';
+export * from './start';
