+++ src/systems/synthesis/service/src/id.ts
@@ -9,10 +9,10 @@ export let ID = createIdGenerator({
   assistant: idType.sorted('ast_'),
   assistantImplementation: idType.sorted('aim_'),
   assistantInstance: idType.sorted('asi_'),
-  assistantConversation: idType.sorted('asc_'),
+  assistantConversation: idType.sorted('acon_'),
   assistantConversationParticipant: idType.sorted('acp_'),
   assistantConversationItem: idType.sorted('aci_'),
-  assistantMessage: idType.sorted('asm_'),
+  assistantMessage: idType.sorted('amsg_'),
   model: idType.sorted('mod_'),
   modelProvider: idType.sorted('mpr_'),
   assistantRequest: idType.sorted('asq_'),
